# docs/fullstack/OPENAPI_TELEMETRY_FHE_V1.yaml
openapi: 3.1.0
info:
  title: NINE65 Telemetry FHE API
  version: 1.0.0
  description: >
    API contract for ciphertext-only telemetry ingestion and homomorphic
    evaluation with NINE65 v5 (`SecureConfig::secure_192()`).
servers:
  - url: https://api.example.com
    description: Production placeholder
  - url: https://staging-api.example.com
    description: Staging placeholder
...
paths:
...
  /v1/fhe/evaluate:
    post:
      tags: [fhe]
      summary: Run homomorphic evaluation against ciphertext inputs
      operationId: evaluateCiphertext
      security:
        - serviceJwt: []
          mutualTLS: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FheEvaluateRequest'
      responses:
        '200':
          description: Evaluation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FheEvaluateResponse'
        '422':
          description: Noise budget policy violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mutualTLS:
      type: mutualTLS
  responses:
    ErrorResponse:
      description: Standard error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    FheEvaluatePolicy:
      type: object
      required: [max_depth, min_noise_budget_bits]
      properties:
        max_depth:
          type: integer
          minimum: 1
          maximum: 64
        min_noise_budget_bits:
          type: integer
          minimum: 1
    FheEvaluateRequest:
      type: object
      required:
        - request_id
        - tenant_id
        - key_id
        - model_version
        - operation
        - ciphertexts
        - policy
      properties:
        request_id:
          type: string
        tenant_id:
          type: string
        key_id:
          type: string
        model_version:
          type: string
        operation:
          type: string
          example: risk_score
        ciphertexts:
          type: array
          minItems: 1
          items:
            type: string
            contentEncoding: base64
        policy:
          $ref: '#/components/schemas/FheEvaluatePolicy'
    FheEvaluateResponse:
      type: object
      required:
        - request_id
        - status
        - result_ciphertext_b64
        - noise_budget_bits
        - latency_ms
        - model_version
      properties:
        request_id:
          type: string
        status:
          type: string
          enum: [completed]
        result_ciphertext_b64:
          type: string
          contentEncoding: base64
        noise_budget_bits:
          type: integer
          minimum: 0
        latency_ms:
          type: integer
          minimum: 0
        model_version:
          type: string
    ErrorDetail:
      type: object
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
          enum:
            - INVALID_PAYLOAD
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT_IDEMPOTENCY
            - RATE_LIMITED
            - NOISE_BUDGET_EXCEEDED
            - INTERNAL_ERROR
            - UPSTREAM_UNAVAILABLE
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'